+++ packages/playground-website/.eslintignore
@@ -1,0 +1,1 @@
+vite.config.ts
