+++ packages/playground-website/.eslintrc.cjs
@@ -1,0 +1,16 @@
+require("@typespec/eslint-config-typespec/patch/modern-module-resolution");
+
+module.exports = {
+  extends: "@typespec/eslint-config-typespec",
+  parserOptions: { tsconfigRootDir: __dirname },
+  rules: {
+    "@typescript-eslint/no-misused-promises": [
+      "error",
+      {
+        checksVoidReturn: {
+          arguments: false, // too much noise when adding async event listeners
+        },
+      },
+    ],
+  },
+};
