+++ packages/playground-website/README.md
@@ -1,0 +1,3 @@
+# TypeSpec Playground
+
+A web app to play with TypeSpec in the browser.
