+++ packages/playground-website/samples/kiosk.tsp
