+++ packages/playground/src/react/error-tab.tsx
