+++ packages/playground/src/react/footer.tsx
