+++ packages/playground/src/react/index.ts
@@ -1,0 +1,1 @@
+export { Playground, PlaygroundProps } from "./playground.jsx";
