+++ packages/playground/src/react/openapi-output.tsx
