+++ packages/playground/src/react/output-tabs.tsx
