+++ rush.json
@@ -168,10 +168,15 @@
       "reviewCategory": "production",
       "shouldPublish": false
     },
-
     {
       "packageName": "@typespec/playground",
       "projectFolder": "packages/playground",
+      "reviewCategory": "production",
+      "shouldPublish": false
+    },
+    {
+      "packageName": "@typespec/playground-website",
+      "projectFolder": "packages/playground-website",
       "reviewCategory": "production",
       "shouldPublish": false
     },
